Refactored for UI integration
