+++ backend/src/main/scala/EntityMatchers/EntityMatch.scala
@@ -1,3 +1,3 @@
 package EntityMatchers
 
-case class EntityMatch(id1:String,id2:String,name1:String,name2:String,similarity:Double)
+case class EntityMatch(id1:String,id2:String,similarity:Double)
