+++ backend/src/main/scala/utilities/Utilities.scala
@@ -1,5 +1,7 @@
 package utilities
 
+import java.io.File
+
 import inputreader.CSVReader
 import org.apache.spark.sql.{DataFrame, Row, SparkSession}
 
@@ -23,5 +25,12 @@ class Utilities extends Serializable {
 }
 
 object Utilities {
+  def deleteRecursively(file: File):Unit={
+    //    val file:File= new File(fileName);
+    if (file.isDirectory)
+      file.listFiles.foreach(deleteRecursively)
+    if (file.exists && !file.delete)
+      throw new Exception(s"Unable to delete ${file.getAbsolutePath}")
+  }
 
 }
